# Try using BitArray for cloud masks