I'm curious if this works on GPUs