@sriharshakandala 2024-02-15
Does this improve performance?